refactoring weather-collector to split up code.
